src/orders_project/requirements.txt
psycopg2==2.7.4
redis==2.10.6
celery==4.1.0
Django==1.11
djangorestframework==3.7.7
requests==2.18.4